requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.1
mysql-connector-python==8.4.0
PyMySQL==1.1.1
sqlparse==0.5.0
typing_extensions==4.11.0
